make or constraint internal
